refactor: 로또 WinningImpl 클래스 싱글톤 변경,패키시 생성
