+++ src/main/java/lotto_model/generator/LottoGenerator.java
@@ -1,4 +1,4 @@
-package lotto_model;
+package lotto_model.generator;
 
 public interface LottoGenerator {
     void printLottoSaved(int inputMoney);
