+++ src/main/java/lotto_model/lotto/Lotto.java
@@ -1,4 +1,4 @@
-package lotto_model;
+package lotto_model.lotto;
 
 import java.util.*;
