+++ src/main/java/lotto_model/lotto/Unit.java
@@ -1,4 +1,4 @@
-package lotto_model;
+package lotto_model.lotto;
 
 public enum Unit {
     UNIT_MONEY(1000),
