+++ src/main/java/lotto_view/LottoInput.java
@@ -1,6 +1,6 @@
 package lotto_view;
 
-import lotto_model.Lotto;
+import lotto_model.lotto.Lotto;
 
 public interface LottoInput {
     int getBonusNumber();
