+++ src/test/java/lotto_view/LottoTest.java
@@ -1,6 +1,6 @@
 package lotto_view;
 
-import lotto_model.Lotto;
+import lotto_model.lotto.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
